Break up primitives.test.ts into multiple files
